refactor(odometer): consumers use union latest-by-date semantics
Delete getLatestOdometer from log.server.ts (max(miles) impl). Wire
reminder.server.ts and vehicle.server.ts to the new odometer.server
helpers (getLatestOdometer / getLatestOdometerByVehicle) so all
callers get the union of logs + manual readings, most-recent-by-date.
Fix the dashboard route import + chip render for the new LatestOdometer
object shape.
